refactor(GAME): Reorganize detection of leave
