+++ front/src/components/router/AppRouter.tsx
@@ -1,15 +1,10 @@
-import InfoBoxContextProvider from '@/context/InfoBoxContext';
 import { BrowserRouter as Router } from 'react-router-dom';
-import InfoBox from '../global/InfoBox';
 import AppRoutes from './AppRoutes';
 
 function AppRouter() {
 	return (
 		<Router>
-			<InfoBoxContextProvider>
-				<InfoBox />
-				<AppRoutes />
-			</InfoBoxContextProvider>
+			<AppRoutes />
 		</Router>
 	);
 }
