+++ front/src/context/InfoBoxContext.tsx
@@ -1,7 +1,4 @@
 import React, { useState } from 'react';
-import { useWebSocket } from 'react-use-websocket/dist/lib/use-websocket';
-import { ws_url as WS_URL } from '@/config.json';
-import { useNavigate } from 'react-router';
 
 export enum InfoType {
 	ERROR = 'error',
@@ -30,37 +27,6 @@ export const InfoBoxContext = React.createContext({
 function InfoBoxContextProvider(props: any) {
 	const [infos, setInfos] = useState<IInfo[]>([]);
 	const [id, setId] = useState(0);
-	const navigate = useNavigate();
-
-	function isGameAbortedEvent(data: any): boolean {
-		return data.event === 'game-aborted';
-	}
-
-	useWebSocket(WS_URL, {
-		share: true,
-		onMessage: ({ data }) => {
-			data = JSON.parse(data);
-			if (isGameAbortedEvent(data)) {
-				if (data.data.reason === 'player-left') {
-					if (data.data.result === 'lose') {
-						addInfo({
-							type: InfoType.ERROR,
-							message: 'You left the game and lost',
-						});
-					} else if (data.data.result === 'win') {
-						addInfo({
-							type: InfoType.SUCCESS,
-							message: 'Your opponent left the game and you won',
-						});
-						navigate('/', { replace: true });
-					}
-				}
-			}
-		},
-		filter: ({ data }) => {
-			return isGameAbortedEvent(JSON.parse(data));
-		},
-	});
 
 	async function removeInfo(id: number) {
 		setInfos((infos) =>
